로그아웃 버튼, 로그인 및 회원가입 이동 버튼 생성 및 동작

+++ src/main/java/com/example/reactdemo/web/model/models/JsonResultApiModel.java
@@ -7,7 +7,7 @@ package com.example.reactdemo.web.model.models;
 */
 public class JsonResultApiModel {
 
-    public short resultCode = 0;
+    public int resultCode = 0;
     public boolean isSuccess = false;
     public String responseCode;
     public Object jsonResult;
